feat: redesign regressor interface
