+++ src/safeds/ml/classical/regression/__init__.py
@@ -5,34 +5,34 @@ from typing import TYPE_CHECKING
 import apipkg
 
 if TYPE_CHECKING:
-    from ._ada_boost import AdaBoostRegressor
+    from ._ada_boost_regressor import AdaBoostRegressor
     from ._arima import ArimaModelRegressor
-    from ._decision_tree import DecisionTreeRegressor
-    from ._elastic_net_regression import ElasticNetRegressor
-    from ._gradient_boosting import GradientBoostingRegressor
-    from ._k_nearest_neighbors import KNearestNeighborsRegressor
-    from ._lasso_regression import LassoRegressor
-    from ._linear_regression import LinearRegressor
-    from ._random_forest import RandomForestRegressor
+    from ._decision_tree_regressor import DecisionTreeRegressor
+    from ._elastic_net_regressor import ElasticNetRegressor
+    from ._gradient_boosting_regressor import GradientBoostingRegressor
+    from ._k_nearest_neighbors_regressor import KNearestNeighborsRegressor
+    from ._lasso_regressor import LassoRegressor
+    from ._linear_regressor import LinearRegressor
+    from ._random_forest_regressor import RandomForestRegressor
     from ._regressor import Regressor
-    from ._ridge_regression import RidgeRegressor
-    from ._support_vector_machine import SupportVectorMachineRegressor
+    from ._ridge_regressor import RidgeRegressor
+    from ._support_vector_regressor import SupportVectorRegressor
 
 apipkg.initpkg(
     __name__,
     {
-        "AdaBoostRegressor": "._ada_boost:AdaBoostRegressor",
+        "AdaBoostRegressor": "._ada_boost_regressor:AdaBoostRegressor",
         "ArimaModelRegressor": "._arima:ArimaModelRegressor",
-        "DecisionTreeRegressor": "._decision_tree:DecisionTreeRegressor",
-        "ElasticNetRegressor": "._elastic_net_regression:ElasticNetRegressor",
-        "GradientBoostingRegressor": "._gradient_boosting:GradientBoostingRegressor",
-        "KNearestNeighborsRegressor": "._k_nearest_neighbors:KNearestNeighborsRegressor",
-        "LassoRegressor": "._lasso_regression:LassoRegressor",
-        "LinearRegressionRegressor": "._linear_regression:LinearRegressionRegressor",
-        "RandomForestRegressor": "._random_forest:RandomForestRegressor",
+        "DecisionTreeRegressor": "._decision_tree_regressor:DecisionTreeRegressor",
+        "ElasticNetRegressor": "._elastic_net_regressor:ElasticNetRegressor",
+        "GradientBoostingRegressor": "._gradient_boosting_regressor:GradientBoostingRegressor",
+        "KNearestNeighborsRegressor": "._k_nearest_neighbors_regressor:KNearestNeighborsRegressor",
+        "LassoRegressor": "._lasso_regressor:LassoRegressor",
+        "LinearRegressor": "._linear_regressor:LinearRegressor",
+        "RandomForestRegressor": "._random_forest_regressor:RandomForestRegressor",
         "Regressor": "._regressor:Regressor",
-        "RidgeRegressor": "._ridge_regression:RidgeRegressor",
-        "SupportVectorMachineRegressor": "._support_vector_machine:SupportVectorMachineRegressor",
+        "RidgeRegressor": "._ridge_regressor:RidgeRegressor",
+        "SupportVectorRegressor": "._support_vector_regressor:SupportVectorRegressor",
     },
 )
 
@@ -48,5 +48,5 @@ __all__ = [
     "RandomForestRegressor",
     "Regressor",
     "RidgeRegressor",
-    "SupportVectorMachineRegressor",
+    "SupportVectorRegressor",
 ]
