+++ src/safeds/ml/classical/regression/_decision_tree_regressor.py
@@ -2,26 +2,24 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from safeds._utils import _get_random_seed, _structural_hash
+from safeds._utils import _structural_hash
 from safeds.exceptions import ClosedBound, OutOfBoundsError
 
-from ._classifier import Classifier
+from ._regressor import Regressor
 
 if TYPE_CHECKING:
-    from sklearn.base import ClassifierMixin
+    from sklearn.base import RegressorMixin
 
     from safeds.data.labeled.containers import TabularDataset
     from safeds.data.tabular.containers import Table
 
 
-class RandomForestClassifier(Classifier):
+class DecisionTreeRegressor(Regressor):
     """
-    Random forest classification.
+    Decision tree regression.
 
     Parameters
     ----------
-    number_of_trees:
-        The number of trees to be used in the random forest. Has to be greater than 0.
     maximum_depth:
         The maximum depth of each tree. If None, the depth is not limited. Has to be greater than 0.
     minimum_number_of_samples_in_leaves:
@@ -29,8 +27,6 @@ class RandomForestClassifier(Classifier):
 
     Raises
     ------
-    OutOfBoundsError
-        If `number_of_trees` is less than 1.
     OutOfBoundsError
         If `maximum_depth` is less than 1.
     OutOfBoundsError
@@ -44,15 +40,12 @@ class RandomForestClassifier(Classifier):
     def __init__(
         self,
         *,
-        number_of_trees: int = 100,
         maximum_depth: int | None = None,
-        minimum_number_of_samples_in_leaves: int = 1,
+        minimum_number_of_samples_in_leaves: int = 5,
     ) -> None:
         super().__init__()
 
         # Validation
-        if number_of_trees < 1:
-            raise OutOfBoundsError(number_of_trees, name="number_of_trees", lower_bound=ClosedBound(1))
         if maximum_depth is not None and maximum_depth < 1:
             raise OutOfBoundsError(maximum_depth, name="maximum_depth", lower_bound=ClosedBound(1))
         if minimum_number_of_samples_in_leaves < 1:
@@ -63,14 +56,12 @@ class RandomForestClassifier(Classifier):
             )
 
         # Hyperparameters
-        self._number_of_trees: int = number_of_trees
         self._maximum_depth: int | None = maximum_depth
         self._minimum_number_of_samples_in_leaves: int = minimum_number_of_samples_in_leaves
 
     def __hash__(self) -> int:
         return _structural_hash(
             super().__hash__(),
-            self._number_of_trees,
             self._maximum_depth,
             self._minimum_number_of_samples_in_leaves,
         )
@@ -80,18 +71,13 @@ class RandomForestClassifier(Classifier):
     # ------------------------------------------------------------------------------------------------------------------
 
     @property
-    def number_of_trees(self) -> int:
-        """The number of trees used in the random forest."""
-        return self._number_of_trees
-
-    @property
     def maximum_depth(self) -> int | None:
-        """The maximum depth of each tree."""
+        """The maximum depth of the tree."""
         return self._maximum_depth
 
     @property
     def minimum_number_of_samples_in_leaves(self) -> int:
-        """The minimum number of samples that must remain in the leaves of each tree."""
+        """The minimum number of samples that must remain in the leaves of the tree."""
         return self._minimum_number_of_samples_in_leaves
 
     # ------------------------------------------------------------------------------------------------------------------
@@ -104,20 +90,16 @@ class RandomForestClassifier(Classifier):
     def _check_additional_predict_preconditions(self, dataset: Table | TabularDataset):
         pass
 
-    def _clone(self) -> RandomForestClassifier:
-        return RandomForestClassifier(
-            number_of_trees=self._number_of_trees,
+    def _clone(self) -> DecisionTreeRegressor:
+        return DecisionTreeRegressor(
             maximum_depth=self._maximum_depth,
             minimum_number_of_samples_in_leaves=self._minimum_number_of_samples_in_leaves,
         )
 
-    def _get_sklearn_model(self) -> ClassifierMixin:
-        from sklearn.ensemble import RandomForestClassifier as SklearnRandomForestClassifier
+    def _get_sklearn_model(self) -> RegressorMixin:
+        from sklearn.tree import DecisionTreeRegressor as SklearnDecisionTreeRegressor
 
-        return SklearnRandomForestClassifier(
-            n_estimators=self._number_of_trees,
+        return SklearnDecisionTreeRegressor(
             max_depth=self._maximum_depth,
             min_samples_leaf=self._minimum_number_of_samples_in_leaves,
-            random_state=_get_random_seed(),
-            n_jobs=-1,
         )
